:recycle: Extract location types, grouping and encoding logic outside of the reporter
