+++ pkg/reporter/sbom/cyclonedx_1_5.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 	"io"
 
+	"github.com/google/osv-scanner/pkg/grouper"
+
 	"github.com/CycloneDX/cyclonedx-go"
 	"github.com/google/osv-scanner/pkg/models"
 )
@@ -14,7 +16,7 @@ func ToCycloneDX15Bom(stderr io.Writer, packageSources []models.PackageSource) *
 	bom.JSONSchema = cycloneDx15Schema
 	bom.SpecVersion = cyclonedx.SpecVersion1_5
 
-	uniquePackages := groupByPackage(packageSources)
+	uniquePackages := grouper.GroupByPURL(packageSources)
 
 	for packageURL, packageDetail := range uniquePackages {
 		component := cyclonedx.Component{}
@@ -26,15 +28,15 @@ func ToCycloneDX15Bom(stderr io.Writer, packageSources []models.PackageSource) *
 		component.Type = componentType
 		component.Evidence = &cyclonedx.Evidence{Occurrences: &occurrences}
 
-		for index, location := range packageDetail.Locations {
-			location, err := createLocationString(location)
+		for index, packageLocations := range packageDetail.Locations {
+			jsonLocation, err := packageLocations.EncodeToJSONString()
 			if err != nil {
-				_, _ = fmt.Fprintf(stderr, "An error occurred when creating the location structure : %v", err.Error())
+				_, _ = fmt.Fprintf(stderr, "An error occurred when creating the jsonLocation structure : %v", err.Error())
 				continue
 			}
 
 			occurrence := cyclonedx.EvidenceOccurrence{
-				Location: location,
+				Location: jsonLocation,
 			}
 			(*component.Evidence.Occurrences)[index] = occurrence
 		}
