+++ pkg/reporter/sbom/models.go
@@ -21,28 +21,6 @@ var SpecVersionToBomCreator = map[CycloneDXVersion]BomCreator{
 
 type BomCreator func(stderr io.Writer, packageSources []models.PackageSource) *cyclonedx.BOM
 
-type packageDetails struct {
-	Name      string
-	Version   string
-	Ecosystem string
-	Locations []PackageLocations
-}
-
-type PackageLocation struct {
-	Filename    string `json:"file_name"`
-	LineStart   int    `json:"line_start"`
-	LineEnd     int    `json:"line_end"`
-	ColumnStart int    `json:"column_start"`
-	ColumnEnd   int    `json:"column_end"`
-}
-
-type PackageLocations struct {
-	Block     *PackageLocation `json:"block"`
-	Namespace *PackageLocation `json:"namespace,omitempty"`
-	Name      *PackageLocation `json:"name,omitempty"`
-	Version   *PackageLocation `json:"version,omitempty"`
-}
-
 const (
 	cycloneDx14Schema = "http://cyclonedx.org/schema/bom-1.4.schema.json"
 	cycloneDx15Schema = "http://cyclonedx.org/schema/bom-1.5.schema.json"
